+++ index.js
@@ -1,2 +1,7 @@
+"use strict";
+
 exports.LayerMixin = require('./LayerMixin');
 exports.Layer = require('./Layer');
+exports.Portal = require('./Portal');
+exports.ContainerMixin = require('./ContainerMixin');
+exports.Container = require('./Container');